+++ src/main/java/com/project/cravehub/controller/user/UserProfileController.java
@@ -1,24 +1,34 @@
 package com.project.cravehub.controller.user;
 
+import com.lowagie.text.DocumentException;
 import com.project.cravehub.dto.AddressDto;
 import com.project.cravehub.dto.ProductDto;
 import com.project.cravehub.dto.UserRegistrationDto;
 import com.project.cravehub.model.admin.Product;
 import com.project.cravehub.model.user.Address;
+import com.project.cravehub.model.user.OrderItem;
+import com.project.cravehub.model.user.PurchaseOrder;
 import com.project.cravehub.model.user.User;
 import com.project.cravehub.repository.AddressRepository;
+import com.project.cravehub.repository.OrderItemRepository;
+import com.project.cravehub.repository.PurchaseOrderRepository;
 import com.project.cravehub.repository.UserRepository;
+import com.project.cravehub.service.InvoiceService;
 import com.project.cravehub.service.addressService.AddressService;
 import com.project.cravehub.service.userservice.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
+import java.io.IOException;
 import java.security.Principal;
-import java.util.List;
-import java.util.Optional;
+import java.util.*;
 
 @Controller
 public class UserProfileController {
@@ -38,10 +48,20 @@ public class UserProfileController {
     @Autowired
     private PasswordEncoder passwordEncoder;
 
+    @Autowired
+    private PurchaseOrderRepository purchaseOrderRepository;
+
+    @Autowired
+    private OrderItemRepository orderItemRepository;
+
+    @Autowired
+    private InvoiceService invoiceService;
+
     @ModelAttribute("user")
     public UserRegistrationDto userRegistrationDto() {
         return  new UserRegistrationDto();
     }
+
     @ModelAttribute("address")
     public AddressDto addressDto() {
         return  new AddressDto();
@@ -81,10 +101,9 @@ public class UserProfileController {
     }
 
     @PostMapping("/addAddress")
-    public String addAddressPost(@ModelAttribute("address") AddressDto addressDto,Principal principal) {
+    public String addAddressPost(@ModelAttribute("address") AddressDto addressDto, Principal principal) {
         String email = principal.getName();
         User user = userRepository.findByEmail(email);
-
         if(user!=null) {
             Address address = new Address();
             address.setUserId(user);
@@ -134,14 +153,57 @@ public class UserProfileController {
         return "redirect:/profile?mismatch";
     }
 
-    @GetMapping("/deleteAddress/{address_id}")
-    public String deleteAddress(@PathVariable("address_id") Integer address_id) {
-        Address deleteAddress = addressService.deleteAddressById(address_id);
+    @PostMapping("/deleteAddress")
+    @ResponseBody
+    public ResponseEntity<String> deleteAddress(@RequestParam("addressId") Integer address_id) {
+        System.out.println("in----------------");
+        if(addressService.isAddressInPurchaseOrder(address_id))
+        {
+            return ResponseEntity.ok("you cant delete this address,you can update.");
+
+        }
+            Address deleteAddress = addressService.deleteAddressById(address_id);
         if(deleteAddress != null)
         {
-            return "redirect:/profile";
+            return ResponseEntity.ok("Address deleted successfully.");
+
         }
-        return "redirect:/profile?addressNotFound";
+        return ResponseEntity.status(500).body("Failed to delete address.");
+    }
+
+    @GetMapping("/orders")
+    public String orders(Principal principal,Model model) {
+        User user = userRepository.findByEmail(principal.getName());
+        model.addAttribute("userName",user.getUserName());
+        List<PurchaseOrder> purchaseOrders = purchaseOrderRepository.findByUser(user);
+        Collections.reverse(purchaseOrders);
+        model.addAttribute("purchaseOrder",purchaseOrders);
+        return "orders";
+    }
+
+    @GetMapping("/cancelOrder/{orderItemId}")
+    public String cancelOrderItem(@PathVariable Integer orderItemId) {
+        Optional<OrderItem> orderItemOptional = orderItemRepository.findById(orderItemId);
+        if(orderItemOptional.isPresent())
+        {
+            OrderItem orderItem = orderItemOptional.get();
+            orderItem.setOrderStatus("cancelled");
+            orderItemRepository.save(orderItem);
+            return "redirect:/orders";
+        }
+        return "redirect:/orders?error";
+    }
+
+    @GetMapping("/downloadInvoice")
+    public ResponseEntity<byte[]> downloadInvoice(@RequestParam("orderItemId") OrderItem orderItem) throws DocumentException, IOException {
+        byte[] pdfBytes = invoiceService.generateInvoice(orderItem);
+        System.out.println("called-------invoice");
+        HttpHeaders headers = new HttpHeaders();
+        headers.setContentType(MediaType.APPLICATION_PDF);
+        headers.setContentDispositionFormData("attachment", "invoice.pdf");
+        headers.setCacheControl("must-revalidate, post-check=0, pre-check=0");
+
+        return ResponseEntity.ok().headers(headers).body(pdfBytes);
     }
 
 }
